# ui/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Refactored free-fs with a subject-match-java style layered architecture.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://192.168.164.18:30067"
    }
  ],
...
  "paths": {
...
    "/files": {
      "get": {
        "summary": "分页查询文件",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultFileRecordDTO"
                }
              }
            }
          }
        },
        "operationId": "search",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "父目录ID",
            "required": false,
            "name": "parentId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "搜索关键词",
            "required": false,
            "name": "keyword"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "name": "fileType"
          },
          {
            "schema": {
              "format": "boolean",
              "type": "boolean"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": true,
            "name": "isDir"
          },
          {
            "schema": {
              "format": "boolean",
              "type": "boolean"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": true,
            "name": "isRecents"
          },
          {
            "schema": {
              "format": "boolean",
              "type": "boolean"
            },
            "in": "query",
            "description": "是否已删除",
            "required": false,
            "example": true,
            "name": "deleted"
          },
          {
            "schema": {
              "format": "boolean",
              "type": "boolean"
            },
            "in": "query",
            "description": "是否收藏",
            "required": false,
            "example": true,
            "name": "favorite"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "排序字段",
            "required": false,
            "name": "sortField"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "排序方向",
            "required": false,
            "name": "sortOrder"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "页索引",
            "required": true,
            "example": 0,
            "name": "pageIndex"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "每页大小",
            "required": true,
            "example": 20,
            "name": "pageSize"
          }
        ]
      }
    },
    "/files/{fileId}": {
      "get": {
        "summary": "查询文件详情",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFileRecordDTO"
                }
              }
            }
          }
        },
        "operationId": "getFileById",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "文件ID",
            "required": true,
            "name": "fileId"
          }
        ]
      }
    },
    "/files/directory": {
      "post": {
        "summary": "创建目录",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDirectoryCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFileRecordDTO"
                }
              }
            }
          }
        },
        "operationId": "createDirectory",
        "parameters": []
      }
    },
    "/files/dirs": {
      "get": {
        "summary": "查询目录列表",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultDirectoryTreeDTO"
                }
              }
            }
          }
        },
        "operationId": "listDirs",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "父目录ID",
            "required": false,
            "name": "parentId"
          }
        ]
      }
    },
    "/files/{fileId}/rename": {
      "put": {
        "summary": "文件重命名",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFileCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "renameFile",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "文件ID",
            "required": true,
            "name": "fileId"
          }
        ]
      }
    },
    "/files/moves": {
      "put": {
        "summary": "文件移动",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveFileCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "moveFile",
        "parameters": []
      }
    },
    "/files/directory/{dirId}/path": {
      "get": {
        "summary": "获取目录层级路径",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultFileRecordDTO"
                }
              }
            }
          }
        },
        "operationId": "getDirectoryTreePath",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "目录ID",
            "required": true,
            "name": "dirId"
          }
        ]
      }
    },
    "/files/url/{fileId}": {
      "get": {
        "summary": "获取文件URL",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultString"
                }
              }
            }
          }
        },
        "operationId": "getFileUrl",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "name": "fileId"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "No comments found.",
            "required": false,
            "example": 0,
            "name": "expireSeconds"
          }
        ]
      }
    },
    "/files/download/{fileId}": {
      "get": {
        "summary": "下载文件",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {}
              }
            }
          }
        },
        "operationId": "downloadFile",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "No comments found.",
            "required": true,
            "name": "fileId"
          }
        ]
      }
    },
    "/files/download/batch": {
      "post": {
        "summary": "批量下载文件(打包为zip)",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {}
              }
            }
          }
        },
        "operationId": "downloadFilesBatch",
        "parameters": []
      }
    },
    "/files/recycle": {
      "get": {
        "summary": "分页查询回收站文件",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultFileRecordDTO"
                }
              }
            }
          }
        },
        "operationId": "listRecycleFiles",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "搜索关键词",
            "required": false,
            "name": "keyword"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "页索引",
            "required": true,
            "example": 0,
            "name": "pageIndex"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "每页大小",
            "required": true,
            "example": 20,
            "name": "pageSize"
          }
        ]
      },
      "delete": {
        "summary": "批量移入回收站",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "moveToRecycle",
        "parameters": []
      }
    },
    "/files/recycle/restore": {
      "put": {
        "summary": "恢复回收站文件",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRecycleCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "restore",
        "parameters": []
      }
    },
...
    "/files/favorites": {
      "post": {
        "summary": "收藏文件",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteFilesCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "favorite",
        "parameters": []
      }
    },
    "/files/unfavorite": {
      "post": {
        "summary": "取消收藏文件",
        "deprecated": false,
        "tags": [
          "FileController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteFilesCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultVoid"
                }
              }
            }
          }
        },
        "operationId": "unfavorite",
        "parameters": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RenameFileCommand": {
        "type": "object",
        "properties": {
          "userId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "fileName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "fileName"
        ]
      },
...
      "SingleResultFileRecordDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/FileRecordDTO"
          }
        }
      },
...
      "PageResultFileRecordDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "totalCount": {
            "format": "int64",
            "description": "总数",
            "type": "number",
            "example": 0
          },
          "pageSize": {
            "format": "int32",
            "description": "页数",
            "type": "integer",
            "example": 0
          },
          "pageIndex": {
            "format": "int32",
            "description": "当前页数",
            "type": "integer",
            "example": 0
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileRecordDTO"
            }
          }
        }
      },
...
    }
  }
}